+++ src/core/models/index.ts
@@ -1,3 +1,2 @@
-export * from "./enums";
-export * from "./types";
-export * from "./interfaces";
+export * from './enums'
+export * from './interfaces'
