+++ src/core/utils/registerComponent.ts
@@ -1,12 +1,12 @@
-import Handlebars, { HelperOptions } from "handlebars";
-import { ClassComponent } from "../models";
+import Handlebars, { HelperOptions } from 'handlebars'
+import { IClassComponent } from '../models'
 
-export function registerComponent(Component: ClassComponent) {
+export function registerComponent(Component: IClassComponent) {
   Handlebars.registerHelper(Component.name, (opts: HelperOptions) => {
-    const { setChild, EEvents, props, state } = opts.data.root;
-    const child = new Component({});
-    
-    console.log("opts", opts.data.root);
-    return "Привет мир";
-  });
+    const { setChild, EEvents, props, state } = opts.data.root
+    const child = new Component({})
+
+    console.log('opts', opts.data.root)
+    return 'Привет мир'
+  })
 }
