+++ src/core/utils/utils.ts
@@ -1,22 +1,22 @@
-import { ISimpleObject } from "../models";
+import { ISimpleObject } from '../models'
 
 export function createProxy<T extends ISimpleObject>(
   data: T,
-  callback?: () => void
+  callback?: () => void,
 ): T {
   return new Proxy(data, {
     set(target: T, key: keyof T, val) {
-      target[key] = val;
-      if (callback) callback();
-      return true;
+      target[key] = val
+      if (callback) callback()
+      return true
     },
-  });
+  })
 }
 
 export function parseEvent(evString: string) {
-  const [event, name] = evString.split(":");
+  const [event, name] = evString.split(':')
   return {
     event: event.trim(),
     name: name.trim(),
-  };
+  }
 }
