+++ src/scripts/main.ts
@@ -1,8 +1,8 @@
-import { TestComponent } from "../components/TestComponent";
-import { App } from "../core/App";
+import { TestComponent } from '../components/TestComponent'
+import { App } from '../core/App'
 
 const app = new App({
   imports: [TestComponent],
-});
+})
 
-app.renderDOM("#root", new TestComponent({}));
+app.renderDOM('#root', new TestComponent({}))
